refactor grid
