+++ src/components/Grid/styles.jsx
@@ -1,0 +1,41 @@
+
+const styles = {
+    fixedToBottom: {
+      position: 'fixed',
+      bottom: '0',
+      right: '0',
+      margin: '10px',
+    },
+    removeStyle: {
+      position: 'absolute',
+      right: '2px',
+      top: 0,
+      cursor: 'pointer',
+    },
+    horizontalHeaderBarStyle: {
+      display: 'inline',
+      whiteSpace: 'nowrap',
+    },
+    iconAlignment: {
+      position: 'fixed',
+      top: '0',
+      right: '0',
+      marginTop: '8px',
+    },
+    gridItemPaperStyle: {
+      height: '100%',
+      width: '100%',
+      margin: '0px',
+      textAlign: 'left',
+      display: 'inline-block',
+    },
+    menuBarStyle: {
+      backgroundColor: 'rgb(0, 188, 212)',
+    },
+    gridItem: {
+      // TODO?
+      // overflow: 'auto',
+    },
+  };
+
+  export default styles;
